+++ src/main/java/domain/Deck.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.card.Card;
 import domain.cardsGenerator.CardsGenerator;
 import domain.participant.Dealer;
 import domain.participant.Participant;
