+++ src/main/java/domain/card/CardNumber.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 import java.util.List;
