+++ src/main/java/domain/cardsGenerator/RandomCardsGenerator.java
@@ -1,8 +1,8 @@
 package domain.cardsGenerator;
 
-import domain.Card;
-import domain.CardNumber;
-import domain.CardShape;
+import domain.card.Card;
+import domain.card.CardNumber;
+import domain.card.CardShape;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
