+++ src/test/java/domain/DealerTest.java
@@ -2,6 +2,10 @@ package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.card.Card;
+import domain.card.CardNumber;
+import domain.card.CardShape;
+import domain.card.Cards;
 import domain.participant.Dealer;
 import domain.participant.Player;
 import java.util.List;
